+++ config/config.go
@@ -31,4 +31,5 @@ func setDefaultValues() {
 	viper.SetDefault(settings.IndentSizeKey, settings.IndentSizeDefault)
 	viper.SetDefault(settings.PreserveRightMarginKey, settings.PreserveRightMarginDefault)
 	viper.SetDefault(settings.HighlightHeadlinesKey, settings.HighlightHeadlinesDefault)
+	viper.SetDefault(settings.RelativeNumberingKey, settings.RelativeNumberingDefault)
 }